+++ src/Client/ManagesSessionTrait.php
@@ -41,6 +41,7 @@ trait ManagesSessionTrait
     {
         $requestId = $this->nextRequestId();
         $request = $this->session->encodeCreateSessionRequest($requestId, $endpointUrl);
+        $this->logger->debug('CreateSession request for {url}', ['url' => $endpointUrl]);
         $this->transport->send($request);
 
         $response = $this->transport->receive();
@@ -48,6 +49,7 @@ trait ManagesSessionTrait
         $decoder = $this->createDecoder($responseBody);
         $sessionResult = $this->session->decodeCreateSessionResponse($decoder);
         $this->authenticationToken = $sessionResult['authenticationToken'];
+        $this->logger->debug('CreateSession response: authToken={token}', ['token' => (string) $this->authenticationToken]);
         $this->dispatch(fn () => new SessionCreated($this, $endpointUrl, $this->authenticationToken));
 
         if (isset($sessionResult['serverNonce'])) {
@@ -81,12 +83,14 @@ trait ManagesSessionTrait
             $userPrivateKey,
             $this->serverNonce,
         );
+        $this->logger->debug('ActivateSession request');
         $this->transport->send($request);
 
         $response = $this->transport->receive();
         $responseBody = $this->unwrapResponse($response);
         $decoder = $this->createDecoder($responseBody);
         $this->session->decodeActivateSessionResponse($decoder);
+        $this->logger->debug('ActivateSession response received');
         $this->dispatch(fn () => new SessionActivated($this, $endpointUrl));
     }
 
@@ -121,6 +125,7 @@ trait ManagesSessionTrait
      */
     private function closeSession(): void
     {
+        $this->logger->debug('CloseSession request');
         $this->dispatch(fn () => new SessionClosed($this));
 
         if ($this->secureChannel !== null && $this->secureChannel->isSecurityActive()) {