+++ src/Client/ManagesSubscriptionsTrait.php
@@ -71,6 +71,7 @@ trait ManagesSubscriptionsTrait
                 $publishingEnabled,
                 $priority,
             );
+            $this->logger->debug('CreateSubscription request (interval={interval}ms)', ['interval' => $publishingInterval]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -78,6 +79,10 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $result = $this->subscriptionService->decodeCreateSubscriptionResponse($decoder);
+            $this->logger->debug('CreateSubscription response: subscriptionId={subId}, revisedInterval={interval}ms', [
+                'subId' => $result->subscriptionId,
+                'interval' => $result->revisedPublishingInterval,
+            ]);
 
             $this->dispatch(fn () => new SubscriptionCreated(
                 $this,
@@ -122,6 +127,10 @@ trait ManagesSubscriptionsTrait
                 $subscriptionId,
                 $monitoredItems,
             );
+            $this->logger->debug('CreateMonitoredItems request: {count} item(s) for subscription {subId}', [
+                'count' => count($monitoredItems),
+                'subId' => $subscriptionId,
+            ]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -129,6 +138,7 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $results = $this->monitoredItemService->decodeCreateMonitoredItemsResponse($decoder);
+            $this->logger->debug('CreateMonitoredItems response: {count} result(s)', ['count' => count($results)]);
 
             foreach ($results as $i => $result) {
                 $itemNodeId = $monitoredItems[$i]['nodeId'] ?? NodeId::numeric(0, ServiceTypeId::NULL);
@@ -180,6 +190,10 @@ trait ManagesSubscriptionsTrait
                 $selectFields,
                 $clientHandle,
             );
+            $this->logger->debug('CreateEventMonitoredItem request for node {nodeId} on subscription {subId}', [
+                'nodeId' => (string) $nodeId,
+                'subId' => $subscriptionId,
+            ]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -187,6 +201,7 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $results = $this->monitoredItemService->decodeCreateMonitoredItemsResponse($decoder);
+            $this->logger->debug('CreateEventMonitoredItem response received');
             $result = $results[0] ?? new MonitoredItemResult(0, 0, 0.0, 0);
 
             $this->dispatch(fn () => new MonitoredItemCreated(
@@ -223,6 +238,10 @@ trait ManagesSubscriptionsTrait
                 $subscriptionId,
                 $monitoredItemIds,
             );
+            $this->logger->debug('DeleteMonitoredItems request: {count} item(s) from subscription {subId}', [
+                'count' => count($monitoredItemIds),
+                'subId' => $subscriptionId,
+            ]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -230,6 +249,7 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $results = $this->monitoredItemService->decodeDeleteMonitoredItemsResponse($decoder);
+            $this->logger->debug('DeleteMonitoredItems response: {count} result(s)', ['count' => count($results)]);
 
             foreach ($results as $i => $statusCode) {
                 $monItemId = $monitoredItemIds[$i] ?? 0;
@@ -264,6 +284,10 @@ trait ManagesSubscriptionsTrait
                 $subscriptionId,
                 $itemsToModify,
             );
+            $this->logger->debug('ModifyMonitoredItems request: {count} item(s) on subscription {subId}', [
+                'count' => count($itemsToModify),
+                'subId' => $subscriptionId,
+            ]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -271,6 +295,7 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $results = $this->monitoredItemService->decodeModifyMonitoredItemsResponse($decoder);
+            $this->logger->debug('ModifyMonitoredItems response: {count} result(s)', ['count' => count($results)]);
 
             foreach ($results as $i => $result) {
                 $monItemId = $itemsToModify[$i]['monitoredItemId'] ?? 0;
@@ -312,6 +337,10 @@ trait ManagesSubscriptionsTrait
                 $linksToAdd,
                 $linksToRemove,
             );
+            $this->logger->debug('SetTriggering request: triggerItem={triggerId} on subscription {subId}', [
+                'triggerId' => $triggeringItemId,
+                'subId' => $subscriptionId,
+            ]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -319,6 +348,7 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $result = $this->monitoredItemService->decodeSetTriggeringResponse($decoder);
+            $this->logger->debug('SetTriggering response received');
 
             $this->dispatch(fn () => new TriggeringConfigured(
                 $this,
@@ -352,6 +382,7 @@ trait ManagesSubscriptionsTrait
                 $this->authenticationToken,
                 [$subscriptionId],
             );
+            $this->logger->debug('DeleteSubscription request: subscriptionId={subId}', ['subId' => $subscriptionId]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -359,6 +390,7 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $results = $this->subscriptionService->decodeDeleteSubscriptionsResponse($decoder);
+            $this->logger->debug('DeleteSubscription response received');
             $statusCode = $results[0] ?? 0;
 
             $this->dispatch(fn () => new SubscriptionDeleted($this, $subscriptionId, $statusCode));
@@ -389,6 +421,7 @@ trait ManagesSubscriptionsTrait
                 $this->authenticationToken,
                 $acknowledgements,
             );
+            $this->logger->debug('Publish request ({ackCount} acknowledgement(s))', ['ackCount' => count($acknowledgements)]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -396,6 +429,10 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $result = $this->publishService->decodePublishResponse($decoder);
+            $this->logger->debug('Publish response: subscriptionId={subId}, {count} notification(s)', [
+                'subId' => $result->subscriptionId,
+                'count' => count($result->notifications),
+            ]);
 
             $this->dispatchPublishEvents($result);
 
@@ -427,6 +464,7 @@ trait ManagesSubscriptionsTrait
                 $subscriptionIds,
                 $sendInitialValues,
             );
+            $this->logger->debug('TransferSubscriptions request: {count} subscription(s)', ['count' => count($subscriptionIds)]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
@@ -434,6 +472,7 @@ trait ManagesSubscriptionsTrait
             $decoder = $this->createDecoder($responseBody);
 
             $results = $this->subscriptionService->decodeTransferSubscriptionsResponse($decoder);
+            $this->logger->debug('TransferSubscriptions response: {count} result(s)', ['count' => count($results)]);
 
             foreach ($results as $i => $transferResult) {
                 $subId = $subscriptionIds[$i] ?? 0;
@@ -466,13 +505,20 @@ trait ManagesSubscriptionsTrait
                 $subscriptionId,
                 $retransmitSequenceNumber,
             );
+            $this->logger->debug('Republish request: subscriptionId={subId}, sequenceNumber={seqNum}', [
+                'subId' => $subscriptionId,
+                'seqNum' => $retransmitSequenceNumber,
+            ]);
             $this->transport->send($request);
 
             $response = $this->transport->receive();
             $responseBody = $this->unwrapResponse($response);
             $decoder = $this->createDecoder($responseBody);
 
-            return $this->subscriptionService->decodeRepublishResponse($decoder);
+            $result = $this->subscriptionService->decodeRepublishResponse($decoder);
+            $this->logger->debug('Republish response received');
+
+            return $result;
         });
     }